Fixed propagation when click child component
Add a stop modifier in child component, so when clicking edit button or delete button on data-table, it should not trigger the click event on table row.
Currently the table row has a click event that trigger onRowClick method. The onRowClick method on ServersTable and ServicesTable will trigger router.push