Fixed undefined issue for tableRow and async await callback with vuex action for createOrUpdateService action.
Only triggering onRowHover method when draggable or showActionsOnHover is enabled.
The vuex action handlers take two parameters including context vuex object and payload. So, the callback will be added as a property of payload object. This ensures the callback will be called only after the patch/post request is done.